// src/api/generated/model/catalogSkuOptionResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Ecommerce Copa API
 * API backend do Ecommerce Copa com autenticação por sessão, catálogo público e CRUD administrativo.
 * OpenAPI spec version: 1.0.0
 */
import type { CatalogSkuOptionResponseAttributes } from "./catalogSkuOptionResponseAttributes.ts";
import type { TagResponse } from "./tagResponse.ts";

export interface CatalogSkuOptionResponse {
	id?: string;
	title?: string;
	description?: string;
	price?: number;
	originalPrice?: number;
	photo?: string;
	stock?: number;
	rating?: number;
	reviewCount?: number;
	/** Valores dos atributos do SKU. */
	attributes?: CatalogSkuOptionResponseAttributes;
	tags?: TagResponse[];
}